+++ docs/conf.py
@@ -24,14 +24,15 @@ import yaml
 #
 # TODO: Update with the official name of your project or product
 
-project = "Documentation starter pack"
+project = "Documentation starter pack with Ulwazi"
 author = "Canonical Ltd."
 
 
 # Sidebar documentation title; best kept reasonably short
 #
 # TODO: To include a version number, add it here (hardcoded or automated).
-#
+version = "beta"
+
 # TODO: To disable the title, set to an empty string.
 
 html_title = project + " documentation"
@@ -60,8 +61,7 @@ html_title = project + " documentation"
 #         -H 'Accept: application/vnd.github.v3.raw' \
 #         https://api.github.com/repos/canonical/<REPO> | jq '.created_at'
 
-copyright = "%s CC-BY-SA, %s" % (datetime.date.today().year, author)
-
+copyright = f"{datetime.date.today().year}"
 
 # Documentation website URL
 #
@@ -111,6 +111,22 @@ html_context = {
     #
     # TODO: To add a tag image, uncomment and update as needed.
     # 'product_tag': '_static/tag.png',
+    #
+    # Inherit project name
+    "project": project, 
+    # Inherit the author value
+    "author": author,
+    # Licensing information
+    # 
+    # TODO: Change your product's license name and a link to its file.
+    # For the name, we recommend using the standard shorthand identifier from
+    # https://spdx.org/licenses
+    # For the URL, link directly to the product's license statement, typically found on
+    # the product's home page or in its GitHub project.
+    "license": {
+        "name": "LGPL-3.0-only",
+        "url": "https://github.com/canonical/ulwazi/blob/main/LICENSE",
+    },
     # Your Discourse instance URL
     #
     # TODO: Change to your Discourse instance URL or leave empty.
@@ -150,7 +166,40 @@ html_context = {
     "display_contributors": False,
 
     # Required for feedback button    
-    'github_issues': 'enabled',
+    "feedback": True,
+    "github_issues": "enabled",
+    "default_source_extension": ".md",
+    "default_edit_url": "https://github.com/canonical/ulwazi/edit/main/docs/index.rst",
+    "default_view_url": "https://github.com/canonical/ulwazi/blob/main/docs/index.rst",
+
+    # Horizontal Nav Menu
+    "company": "Canonical",
+    "link1_URL": "https://snapcraft.io/",
+    "link1_name": "First optional link",
+    "link2_URL": "https://snapcraft.io/",
+    "link2_name": "Second optional link",
+
+    # Canonical Product menu
+    # Uncomment if you need a product menu added on the top of every page
+    "add_product_menu": True,
+    
+    # Main Horizontal menu
+    # "is_docs": False, # Purpose unknown
+    "logo_link_URL": "https://documentation.ubuntu.com",
+    "logo_img_URL": "https://assets.ubuntu.com/v1/82818827-CoF_white.svg",
+    "logo_title": "Canonical",
+
+    # TODO: Customize the footer.
+    "footer": {
+        # Whether to add the product name as the first entry.
+        "product": True,
+        # Whether to add the license as the second entry.
+        "license": True,
+        # List your footer entries. Accepts HTML tags.
+        "entries": [
+            '<a class="js-revoke-cookie-manager" href="#tracker-settings">Manage your tracker settings</a>',
+        ]
+    }
 }
 
 html_extra_path = []
@@ -177,6 +226,9 @@ if os.getenv("OPENAPI", ""):
 #       uncomment and update as needed.
 
 # slug = ''
+
+# Limit the number of levels for Table of contents
+localtoc_max_depth = 3
 
 #######################
 # Sitemap configuration: https://sphinx-sitemap.readthedocs.io/
@@ -212,6 +264,7 @@ sitemap_excludes = [
 #html_static_path = ["_static"]
 #templates_path = ["_templates"]
 
+html_theme = "ulwazi"
 
 #############
 # Redirects #
@@ -266,8 +319,12 @@ linkcheck_retries = 3
 # NOTE: By default, the following MyST extensions are enabled:
 #       substitution, deflist, linkify
 
-# myst_enable_extensions = set()
-
+myst_enable_extensions = {
+    "colon_fence",
+    "deflist",
+    "tasklist",
+    "substitution",
+}
 
 # Custom Sphinx extensions; see
 # https://www.sphinx-doc.org/en/master/usage/extensions/index.html
@@ -275,7 +332,7 @@ linkcheck_retries = 3
 # NOTE: The canonical_sphinx extension is required for the starter pack.
 
 extensions = [
-    "canonical_sphinx",
+    "canonical_sphinx_config",
     "notfound.extension",
     "sphinx_design",
     "sphinx_reredirects",
@@ -294,6 +351,9 @@ extensions = [
     "sphinx_last_updated_by_git",
     "sphinx.ext.intersphinx",
     "sphinx_sitemap",
+    "myst_parser",
+    "ulwazi",
+    "sphinx_modern_pdf_style",
 ]
 
 # Excludes files or directories from processing
@@ -311,6 +371,10 @@ exclude_patterns = [
 
 # html_js_files = []
 
+# Syntax highlighting settings
+
+highlight_language = "none" # default
+pygments_style = "autumn" # see https://pygments.org/styles for more
 
 # Specifies a reST snippet to be appended to each .rst file
 
@@ -371,3 +435,8 @@ intersphinx_mapping = {
     'starter-pack': ('https://canonical-example-product-documentation.readthedocs-hosted.com/en/latest', None),
     'sphinxcontrib-mermaid': ('https://sphinxcontrib-mermaid-demo.readthedocs.io/en/latest', None)
 }
+
+[...]
+# PDF
+
+set_modern_pdf_config = True